Add a matrix class implementation to the project
Add a class representing a matrix to the project files. Its contents are
saved using a two-dimensional array, and various operations such as
multiplication and exponentiation by squaring are supported. Include
unit tests to test the implemented operations. This class will be used
for another fibonacci algorithm calculating its results in logarithmic
time.