Add meter classes
Add two classes, WallMeter and CycleMeter, which can be used to measure
the execution time of code, both of which are subclasses of an abstract
class called Meter. The WallMeter class utilizes the StopWatch type from
std.datetime to measure the wall clock time in nanoseconds, whereas
CycleMeter relies on the rdtsc assembly instruction to measure the CPU
cycle count. Both of them will be used to measure the efficiency of
various algorithms.