Implement the Heapsort algorithm
Add functions used to operate on heaps. These functions support
returning parent, left child, and right child to a given node, as well
as maintining the heap property for a given node, and making a heap of a
given array. These functions are used by the Heapsort algorithm to sort
a given vector in place using O(n log n) time.