Implement measurements for the sorting algorithms
Add another measurement class SortMeasurement used for taking various
measurements of the sorting algorithms. It contains a vector of data on
whose subsets the measurement operations are performed, as well as
functions for setting specific arrangements (ascending, descending,
random, random with repeated elements) for it. It also holds a function
measureForAll() which can be used to perform measurements of all of the
available vector arrangements within a single step.