Change graph representation to int[vert][vert]
Discard AdjElement in favor to representing a graph using an associative
array of associative arrays, whose elements would equal the weight of an
edge. A representation of an adjacency list using AdjElement gains a
little bit of code readability, however it brings the disadvantage of
having to handle duplicates upon isertion or other operations on the
graph, as well as some complexity deleting them from the array (which
would need to be rearranged). Using an associative array duplicates can,
by definition, not occur. Furthermore getting the weight of an edge
becomes as simple as just accessing it by its index, and deletion can be
realized using the remove() function. Change the functions accordingly.