Implement graph reduction
Implement an algorithm to reduce a given graph. This algorithm
yields a graph in whose adjacency list any edge is represented exactly
one time. Since the graphs operated on are undirected, using this method
a more compact representation of the graph can be achieved by deducing
the missing direction. Modify the methods for insertion and deletion of
edges to optionally operate one-directional, adding an additional
parameter to them. Rewrite the unit tests to test all implementations.