Implement restore, construct, and safe edges
Implement another function, called restore, to restore a previously
reduced graph. Implement a constructing function used to generate a
graph from a given set of vertices and edges. Implement a function
finding a safe edge for a given subset of a graph. Safe edges are edges
whose insertion does not cause a cycle, so this can be utilized to
generate spanning trees. Expand the unit tests accordingly.