Implement generic MST algorithm
Implement a simple algorithm used to generate a minimum spanning tree.
This algorithm makes use of the safeEdge() function by repeatedly
inserting safe edges, starting from the first node of the graph. This is
a simple method for generating minimum spanning trees, however it will
not work if the given graph consists of more than one component.