Generalize graph definitions
Make the definition of a graph generic, having a template parameter for
each the type of the vertex and the type of the edge. This way, graphs
consisting of any data type can be constructed. Apart from the simple
case of generating graphs holding other data types, for instance, a
depth-first search algorithm could too be implemented in a way that the
types of found edges are stored along with their weight. If no
parameters are specified, a default type will be used. Modify all
functions and unit tests to comply with the new definition.