Implement function to check for tree property
Implement a function to check whether a given graph is a tree. This
algorithm makes use of the previously implemented depth-first search
algorithm by looking for back edges. If none are found, 'true' is being
returned. Extend the unit test and use a more complex test case to check
whether the algorithms are correct.