Implement MST measurement
Implement a measurement used to measure the performance of MST
algorithms. The MSTMeasurement works by generating random graphs,
inserting new vertices and connecting them with a defined amount of
randomly chosen other vertices as the graph grows. Multiple output files
are generated depending on the degree of the inserted vertices.
Currently, only the genericMST algorithm is being measured as no other
algorithms have yet been implemented, however this will be changed in
the future.