Implement disjoint set forest data structure
Implement a disjoint set forest (DSF) and add it to the project. Such
data structure can be used to group various kinds of data into disjoint
sets, allowing to quickly check whether two given elements belong to the
same set, as well as to unify two given sets. The sets of the forest are
implemented in a tree-like structure, with each element pointing to its
parent, and the root (the representative of a set) pointing to itself.
Two heuristics (union by rank, path compression) are used to gain a
near-linear runtime for operations on the data structure.