Implement Kruskal's algorithm
Implement Kruskal's algorithm for finding a minimum spanning tree in a
given graph. Kruskal's algorithm is a specialization of the previously
implemented genericMST algorithm. A spanning tree is generated by
grouping all edges of the graph in a priority queue with ascending
weight, and adding them to the graph if no cycle is caused. To determine
whether inserting an edge would cause a cycle, all connected components
of the current graph are tracked using a Disjoint Set Forest. A cycle
would be caused if and only if both vertices of an edge belong to the
same set.